kbdsensor: fix system_server high load issue
After the change 1f9a8a2e, if the sensor is invalid, the polling method
will return immediately without any way to throttle the number of
executions of the main sensor thread loop. That causes CPU high load.

Indeed the HAL should not set the pointer of struct hw_device_t if the
initialization fails. Then the SensorService won't enable and poll the
sensor device.

Fixes: 1f9a8a2e (kbdsensor: update HAL version to 1_3)
Tested-by: Michael Goffioul <michael.goffioul@gmail.com>