kbdsensor: fix parsing of input device name
Originally we use %s in sscanf to get the device name from the property.
However, %s only matches a sequence of non-white-space characters. That
means if the name contains space characters, it can't be parsed correctly.

The patch uses strsep to extract the substring before the first delimiter
as the device name.