Fix kbd sensor
The native SensorService expects individual sensor to send
FLUSH_COMPLETE metadata event in response to a flush command. Until the
event is triggered, no event from that sensor is dispatched to
listeners. By making the kbd sensor to return -EINVAL, the native
service will not wait for the FLUSH_COMPLETE event, and events will be
dispatched right away.